+++ ports/stm32/spi.h
@@ -34,6 +34,16 @@ typedef struct _spi_t {
     const dma_descr_t *rx_dma_descr;
 } spi_t;
 
+typedef struct _pyb_spi_obj_t {
+    mp_obj_base_t base;
+    const spi_t *spi;
+} pyb_spi_obj_t;
+
+typedef struct _machine_hard_spi_obj_t {
+    mp_obj_base_t base;
+    const spi_t *spi;
+} machine_hard_spi_obj_t;
+
 extern SPI_HandleTypeDef SPIHandle1;
 extern SPI_HandleTypeDef SPIHandle2;
 extern SPI_HandleTypeDef SPIHandle3;
@@ -47,8 +57,19 @@ extern const mp_obj_type_t pyb_spi_type;
 extern const mp_obj_type_t machine_soft_spi_type;
 extern const mp_obj_type_t machine_hard_spi_type;
 
+// A transfer of "len" bytes should take len*8*1000/baudrate milliseconds.
+// To simplify the calculation we assume the baudrate is never less than 8kHz
+// and use that value for the baudrate in the formula, plus a small constant.
+#define SPI_TRANSFER_TIMEOUT(len) ((len) + 100)
+
 void spi_init0(void);
 void spi_init(const spi_t *spi, bool enable_nss_pin);
+void spi_deinit(const spi_t *spi_obj);
+int spi_find_index(mp_obj_t id);
+void spi_set_params(const spi_t *spi_obj, uint32_t prescale, int32_t baudrate,
+    int32_t polarity, int32_t phase, int32_t bits, int32_t firstbit);
+void spi_transfer(const spi_t *self, size_t len, const uint8_t *src, uint8_t *dest, uint32_t timeout);
+void spi_print(const mp_print_t *print, const spi_t *spi_obj, bool legacy);
 const spi_t *spi_from_mp_obj(mp_obj_t o);
 
 #endif // MICROPY_INCLUDED_STM32_SPI_H
